% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exrcise-package.R
\docType{package}
\name{exrcise-package}
\alias{exrcise-package}
\title{exrcise: A package to selectively excise code chunks for exercises.}
\description{
The exrcise package selectively removes the contents of code chunks from
Rmarkdown files to produce new notebooks suitable for live coding instruction
or exercises, while preserving the original file as a reference or answer key.
}
