% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exrcise.R
\docType{package}
\name{exrcise}
\alias{exrcise}
\title{exrcise: A package to selectively excise code chunks for exercises.}
\usage{
exrcise(infile, outfile, replace_flags, replacement = "")
}
\arguments{
\item{infile}{a file string or file handle of a .Rmd file
from which code chunks will be removed}

\item{outfile}{a file string or file handle of a .Rmd file
to write the excised file}

\item{replace_flags}{a character vector of code chunk options
that if set to TRUE will have their code replaced.}

\item{replacement}{a character vector that code chunks will be
replaced with, one element per line.}
}
\value{
a character vector of the output lines, invisibly.
}
\examples{
exrcise(infile, outfile,
        replace_flags = "solution",
        replacement = "### Your Code Here")
}
