% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_workout.R
\name{generate_workout}
\alias{generate_workout}
\title{Generate a random workout}
\usage{
generate_workout(warmupREPS, workoutREPS)
}
\arguments{
\item{warmupREPS}{input number of desired warm-ups, numeric (i.e., 3). Maximum is 15.}

\item{workoutREPS}{input number of desiered workouts, numeric (i.e., 4). Maximum is 32.}
}
\value{
A data frame that lists warm-ups and workouts (user defined), and splits moves that require left and right sides into a LEFT and RIGHT size.

Make sure to set equal to a new data frame to save this workout. Then use this new data frame to run the working using the \code{run_workout()} function.
}
\description{
Returns a data frame with a list of warm-up and workout exercises. To be used in \code{run_workout()}.
}
\examples{
myworkout <- generate_workout(4, 8)
# Output will include 4 warm-ups and 8 workouts

myworkout <- generate_workout(3, 5)
myworkout

# Example output:
# Workout
# 1           lateral / side lunges (alternate sides)
# 2                                      butt kickers
# 3                                       jump squats
# 4                                 single-leg squats
# 5                                 single-leg squats
# 6 step-up / box jump (use stair, a chair, or couch)
# 7                                       plank jacks
# 8                                        plank hold
# 9                                            squats
# Type Sides
# 1 warm-up  both
# 2 warm-up  both
# 3 warm-up  both
# 4    main right
# 5    main  left
# 6    main  both
# 7    main  both
# 8    main  both
# 9    main  both


}
