% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tabata_nosound.R
\name{run_tabata_workout_nosound}
\alias{run_tabata_workout_nosound}
\title{Run a tabata workout with no sound}
\usage{
run_tabata_workout_nosound(rest, set)
}
\arguments{
\item{rest}{Indicate the number of seconds in between sets. Numeric, recommend 10 (10 seconds).}

\item{set}{Indicate the length for each of the 4 exercises in seconds, recommend 30 (30 seconds).}
}
\value{
After executing this command, the console will turn into a countdown timer to run your workout. Sounds will not play betwen each exercise or transition.

The four exercises randomly selected will be counted down at the user specified set length (in seconds). Then the four exercises will be repeated. 4 exercises x 2 sets.
}
\description{
Selects a random subset of 4 workout exercises and starts the countdown timer to run in a 'tabata' format, where the 4 exercises will be repeated once. No sound or alarm.
}
\examples{

run_tabata_workout_nosound(10, 30)

# Will start a workout set of 4 exercises
## that will be repeated once (4 x2).
## Workouts will run for 30 seconds with
## 10 second transitions. No sound will be played.



}
