% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exrcise.R
\name{strip_chunks}
\alias{strip_chunks}
\title{Title}
\usage{
strip_chunks(block, chunk_begin, strip_args, replacement = "")
}
\arguments{
\item{block}{}

\item{chunk_begin}{}

\item{strip_args}{}

\item{replacement}{}
}
\value{

}
\description{
Title
}
