% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exrcise.R
\name{strip_document}
\alias{strip_document}
\title{Strip an rmarkdown document of solution code}
\usage{
strip_document(lines, strip_args, replacement = "")
}
\arguments{
\item{lines}{}

\item{strip_args}{}

\item{replacement}{}
}
\value{
A vector of lines that forms a new .Rmd document
}
\description{
Strip an rmarkdown document of solution code
}
