% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-workout.R
\docType{data}
\name{workout_list}
\alias{workout_list}
\title{List of workout moves}
\format{
A data frame with 47 rows and 3 columns:
\describe{
\item{Workout}{Name of workout or warm-up}
\item{Type}{Either 'warm-up' to specify a warm-up or 'main' to designate a main workout.}
\item{Sides}{Indicates if the warm-up or workout requires reps that are left and right sided, specified by 'left,right'. Otherwise, 'both'}
...
}
}
\usage{
workout_list
}
\description{
A dataframe containing all warm-ups and workouts. This list will be selected from for the package functions. Columns provide additional information, including if the move requires a repeat on the left and right side.
}
\keyword{datasets}
