% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exrcise.R
\docType{package}
\name{exrcise}
\alias{exrcise}
\title{exrcise: A package to selectively excise code chunks for exercises.}
\usage{
exrcise(infile, outfile, replace_flags, replacement = "", comment = TRUE)
}
\arguments{
\item{infile}{A file path or file handle of an .Rmd file
from which code chunks will be removed}

\item{outfile}{A file path or file handle to which to
write the .Rmd file with replacements.}

\item{replace_flags}{A character vector of code chunk options
that if set to TRUE will have their code replaced.}

\item{replacement}{A character vector that code chunks will be
replaced with, one element per line.}

\item{comment}{TRUE or FALSE; whether to include comments.
Default is TRUE}
}
\value{
A character vector of the output lines, invisibly.
}
\examples{
\dontrun{
exrcise(infile, outfile,
    replace_flags = "live",
    replacement = "")

exrcise(infile, outfile,
    replace_flags = "solution",
    replacement = "### Your Code Here",
    comment = FALSE)
}

}
