% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exrcise.R
\name{exrcise}
\alias{exrcise}
\title{Read an Rmd file and write a version with selective code chunks replaced}
\usage{
exrcise(infile, outfile, replace_flags, replacement = "")
}
\arguments{
\item{replacement}{}
}
\value{

}
\description{
Read an Rmd file and write a version with selective code chunks replaced
}
\examples{
exrcise(infile, outfile,
        replace_flags = "solution",
        replacement = "### Your Code Here")
}
