% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_code_chunks.R
\name{replace_block}
\alias{replace_block}
\title{Title}
\usage{
replace_block(block, replace_flags, replacement = "")
}
\arguments{
\item{block}{}

\item{replace_flags}{}

\item{replacement}{}
}
\value{

}
\description{
Title
}
