% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_code_chunks.R
\name{replace_block}
\alias{replace_block}
\title{Conditionally replace code in a block of lines}
\usage{
replace_block(block, replace_flags, replacement, comment)
}
\arguments{
\item{comment}{}
}
\value{

}
\description{
Conditionally replace code in a block of lines
}
