% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exrcise.R
\name{replace_code_chunks}
\alias{replace_code_chunks}
\title{Strip an rmarkdown represented as a vector of lines of solution code}
\usage{
replace_code_chunks(lines, replace_flags, replacement = "")
}
\arguments{
\item{lines}{}

\item{replace_flags}{}

\item{replacement}{}
}
\value{
A vector of lines that forms a new .Rmd document
}
\description{
Strip an rmarkdown represented as a vector of lines of solution code
}
\examples{
replace_code_chunks(lines, replace_flags = "solution", replacement = "### Your Code Here")
}
