% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_code_chunks.R
\name{replace_code_chunks}
\alias{replace_code_chunks}
\title{Replace code from an rmarkdown document as a vector of lines}
\usage{
replace_code_chunks(lines, replace_flags, replacement = "")
}
\arguments{
\item{lines}{}

\item{replace_flags}{}

\item{replacement}{}
}
\value{
A vector of lines that forms a new .Rmd document
}
\description{
Replace code from an rmarkdown document as a vector of lines
}
\examples{
replace_code_chunks(lines, 
    replace_flags = "solution", 
    replacement = "### Your Code Here")
}
